fix: memory leak in react comps
