+++ src/main/modules/game/index.ts
@@ -1,3 +1,3 @@
 /* eslint-disable import/prefer-default-export */
-export { default as GameBoardInstructionsParser } from './gameBoardInstructionsParser';
-export { default as GameBoardUDPClient } from './gameBoardUDPClient';
+export { default as GameBoardInstructionsParser } from './instructions/gameBoardInstructionsParser';
+export { default as GameBoardUDPClient } from './data/gameBoardUDPClient';
