+++ src/main/modules/network/index.ts
@@ -1,2 +1,2 @@
 /* eslint-disable import/prefer-default-export */
-export { default as getCurrentConnection } from './currentConnections';
+export { default as getCurrentConnection } from './windows/currentConnections';
