+++ src/renderer/components/board/Points.tsx
@@ -1,4 +1,4 @@
-import { FC, useState } from 'react';
+import { FC, useEffect, useState } from 'react';
 import { EGameBoardDisplayChannels } from 'shared/enums';
 
 interface IPoints {
@@ -6,14 +6,16 @@ interface IPoints {
 }
 const Points: FC<IPoints> = ({ channel }) => {
   const [points, setPoints] = useState('0');
-  window.electron.ipcRenderer.on(channel, (arg) => {
-    try {
-      const pointsData = arg as string;
-      if (pointsData !== points) setPoints(pointsData);
-    } catch (err) {
-      console.error(err);
-    }
-  });
+  useEffect(() => {
+    window.electron.ipcRenderer.on(channel, (arg) => {
+      try {
+        const pointsData = arg as string;
+        if (pointsData !== points) setPoints(pointsData);
+      } catch (err) {
+        console.error(err);
+      }
+    });
+  }, []);
 
   return (
     <div className="text-6xl font-medium flex items-center justify-center p-2">
