+++ src/renderer/components/board/ShotClock.tsx
@@ -1,4 +1,4 @@
-import { FC, useState } from 'react';
+import { FC, useEffect, useState } from 'react';
 import { EGameBoardDisplayChannels } from 'shared/enums';
 
 const ShotClock: FC = () => {
@@ -7,21 +7,23 @@ const ShotClock: FC = () => {
     if (number >= start && number <= end) return true;
     return false;
   };
-  window.electron.ipcRenderer.on(
-    EGameBoardDisplayChannels.additionalClockChannel,
-    (arg) => {
-      try {
-        const secondsData = arg as string;
-        if (secondsData !== seconds) {
-          if (inRange(+secondsData, 0, 24)) {
-            setSeconds(secondsData);
-          } else setSeconds('--');
+  useEffect(() => {
+    window.electron.ipcRenderer.on(
+      EGameBoardDisplayChannels.additionalClockChannel,
+      (arg) => {
+        try {
+          const secondsData = arg as string;
+          if (secondsData !== seconds) {
+            if (inRange(+secondsData, 0, 24)) {
+              setSeconds(secondsData);
+            } else setSeconds('--');
+          }
+        } catch (err) {
+          console.error(err);
         }
-      } catch (err) {
-        console.error(err);
       }
-    }
-  );
+    );
+  }, []);
 
   return (
     <div className="text-4xl flex items-center justify-center py-2">
