+++ src/renderer/components/board/Team.tsx
