+++ src/shared/utils.ts
@@ -1,0 +1,9 @@
+/* eslint-disable import/prefer-default-export */
+export const inRange = (
+  number: number,
+  start: number,
+  end: number
+): boolean => {
+  if (number >= start && number <= end) return true;
+  return false;
+};
